+++ src/compas_timber/fabrication/double_cut.py
@@ -1,5 +1,4 @@
 import math
-from collections import OrderedDict
 
 from compas.geometry import Frame
 from compas.geometry import Line
@@ -12,7 +11,6 @@ from compas.geometry import angle_vectors
 from compas.geometry import dot_vectors
 from compas.geometry import intersection_line_plane
 from compas.geometry import intersection_plane_plane
-from compas.tolerance import TOL
 
 from compas_timber.errors import FeatureApplicationError
 from compas_timber.utils import intersection_line_beam_param
@@ -435,24 +433,17 @@ class DoubleCutParams(BTLxProcessingParams):
         # type: (DoubleCut) -> None
         super(DoubleCutParams, self).__init__(instance)
 
-    def as_dict(self):
-        """Returns the parameters of the Double Cut feature as a dictionary.
-
-        Returns
-        -------
-        dict
-            The parameters of the Double Cut feature as a dictionary.
-        """
-        # type: () -> OrderedDict
-        result = OrderedDict()
-        result["Orientation"] = self._instance.orientation
-        result["StartX"] = "{:.{prec}f}".format(float(self._instance.start_x), prec=TOL.precision)
-        result["StartY"] = "{:.{prec}f}".format(float(self._instance.start_y), prec=TOL.precision)
-        result["Angle1"] = "{:.{prec}f}".format(float(self._instance.angle_1), prec=TOL.precision)
-        result["Inclination1"] = "{:.{prec}f}".format(float(self._instance.inclination_1), prec=TOL.precision)
-        result["Angle2"] = "{:.{prec}f}".format(float(self._instance.angle_2), prec=TOL.precision)
-        result["Inclination2"] = "{:.{prec}f}".format(float(self._instance.inclination_2), prec=TOL.precision)
-        return result
+    @property
+    def attribute_map(self):
+        return {
+            "Orientation": "orientation",
+            "StartX": "start_x",
+            "StartY": "start_y",
+            "Angle1": "angle_1",
+            "Inclination1": "inclination_1",
+            "Angle2": "angle_2",
+            "Inclination2": "inclination_2",
+        }
 
 
 class DoubleCutProxy(object):