+++ src/compas_timber/fabrication/lap.py
@@ -1,5 +1,4 @@
 import math
-from collections import OrderedDict
 
 from compas.datastructures import Mesh
 from compas.geometry import Brep
@@ -22,11 +21,11 @@ from compas.geometry import is_point_behind_plane
 from compas.tolerance import TOL
 
 from compas_timber.errors import FeatureApplicationError
+from compas_timber.fabrication.btlx import MachiningLimits
 from compas_timber.utils import planar_surface_point_at
 
 from .btlx import BTLxProcessing
 from .btlx import BTLxProcessingParams
-from .btlx import MachiningLimits
 from .btlx import OrientationType
 
 
@@ -61,7 +60,7 @@ class Lap(BTLxProcessing):
         The lead inclination is parallel to the beam axis.
     lead_inclination : float
         The lead inclination of the cut. 0.1 < lead_inclination < 179.9.
-    machining_limits : dict, optional
+    machining_limits : :class:`~compas_timber.fabrication.btlx.MachiningLimits` or dict, optional
         The machining limits for the cut. Default is None
 
     """
@@ -84,7 +83,7 @@ class Lap(BTLxProcessing):
         data["lead_angle"] = self.lead_angle
         data["lead_inclination_parallel"] = self.lead_inclination_parallel
         data["lead_inclination"] = self.lead_inclination
-        data["machining_limits"] = self.machining_limits
+        data["machining_limits"] = self.machining_limits.limits
         return data
 
     # fmt: off
@@ -281,15 +280,14 @@ class Lap(BTLxProcessing):
 
     @machining_limits.setter
     def machining_limits(self, machining_limits):
-        if not isinstance(machining_limits, dict):
-            raise ValueError("Machining limits must be a dictionary.")
-        for key, value in machining_limits.items():
-            if key not in MachiningLimits.EXPECTED_KEYS:
-                raise ValueError("The key must be one of the following: ", {self.EXPECTED_KEYS})
-            if not isinstance(value, bool):
-                raise ValueError("The values must be a boolean.")
-        self._machining_limits = machining_limits
-
+        if isinstance(machining_limits, MachiningLimits):
+            self._machining_limits = machining_limits
+        elif isinstance(machining_limits, dict):
+            self._machining_limits = MachiningLimits.from_dict(machining_limits)
+        elif machining_limits is None:
+            self._machining_limits = MachiningLimits()
+        else:
+            raise ValueError("Invalid machining limits")
 
     ########################################################################
     # Alternative constructors
@@ -369,7 +367,7 @@ class Lap(BTLxProcessing):
                    length=length,
                    width=width,
                    depth=depth,
-                   machining_limits=machining_limits.limits,
+                   machining_limits=machining_limits,
                    ref_side_index=ref_side_index)
 
     @classmethod
@@ -382,7 +380,7 @@ class Lap(BTLxProcessing):
             The volume of the lap. Must have 6 faces.
         beam : :class:`~compas_timber.elements.Beam`
             The beam that is cut by this instance.
-        machining_limits : dict, optional
+        machining_limits : :class:`~compas_timber.fabrication.MachiningLimits` or dict
             The machining limits for the cut. Default is None.
         ref_side_index : int, optional
             The index of the reference side of the element. Default is 0.
@@ -472,10 +470,7 @@ class Lap(BTLxProcessing):
             lead_inclination = angle_vectors_signed(front_plane.normal, bottom_plane.normal, ref_side.xaxis, deg=True)
 
         # define machining limits
-        if machining_limits:
-            if not isinstance(machining_limits, dict):
-                raise ValueError("machining_limits must be a dictionary.")
-        else:
+        if not machining_limits:
             machining_limits = cls._define_machining_limits(planes, beam, ref_side_index)
 
         return cls(
@@ -637,7 +632,7 @@ class Lap(BTLxProcessing):
         machining_limits.face_limited_front = is_point_behind_plane(front_plane.point, front_side)
         machining_limits.face_limited_back = is_point_behind_plane(back_plane.point, back_side)
         machining_limits.face_limited_bottom = is_point_behind_plane(bottom_plane.point, opp_side)
-        return machining_limits.limits
+        return machining_limits
 
 
     ########################################################################
@@ -749,26 +744,26 @@ class Lap(BTLxProcessing):
         top_frame = beam.ref_sides[self.ref_side_index] # top should always be unlimited
         top_frame.translate(top_frame.normal * tol)
 
-        if self.machining_limits["FaceLimitedEnd"]:
+        if self.machining_limits.face_limited_end:
             end_frame = start_frame.translated(-start_frame.normal * self.length)
             end_frame.yaxis = -end_frame.yaxis
         else:
             end_frame = beam.ref_sides[5]
 
-        if self.machining_limits["FaceLimitedBottom"]:
+        if self.machining_limits.face_limited_bottom:
             bottom_frame = Frame(start_frame.point, start_frame.zaxis, start_frame.yaxis)
             angle = angle_vectors_signed(top_frame.xaxis, -start_frame.xaxis, top_frame.yaxis)
             bottom_frame = bottom_frame.translated(bottom_frame.zaxis * (self.depth/math.sin(angle)))
         else:
             bottom_frame = beam.opp_side(self.ref_side_index)
 
-        if self.machining_limits["FaceLimitedFront"]:
+        if self.machining_limits.face_limited_front:
             front_frame = bottom_frame.rotated(math.radians(self.lead_angle), bottom_frame.xaxis, point=bottom_frame.point)
         else:
             front_frame = beam.front_side(self.ref_side_index)
             front_frame.translate(front_frame.normal * tol)
 
-        if self.machining_limits["FaceLimitedBack"]:
+        if self.machining_limits.face_limited_back:
             back_frame = front_frame.translated(-front_frame.zaxis * self.width)
             back_frame.xaxis = -back_frame.xaxis
         else:
@@ -857,31 +852,24 @@ class LapParams(BTLxProcessingParams):
         # type: (Lap) -> None
         super(LapParams, self).__init__(instance)
 
-    def as_dict(self):
-        """Returns the parameters of the Lap feature as a dictionary.
-
-        Returns
-        -------
-        dict
-            The parameters of the Lap feature as a dictionary.
-        """
-        # type: () -> OrderedDict
-        result = OrderedDict()
-        result["Orientation"] = self._instance.orientation
-        result["StartX"] = "{:.{prec}f}".format(float(self._instance.start_x), prec=TOL.precision)
-        result["StartY"] = "{:.{prec}f}".format(float(self._instance.start_y), prec=TOL.precision)
-        result["Angle"] = "{:.{prec}f}".format(float(self._instance.angle), prec=TOL.precision)
-        result["Inclination"] = "{:.{prec}f}".format(float(self._instance.inclination), prec=TOL.precision)
-        result["Slope"] = "{:.{prec}f}".format(float(self._instance.slope), prec=TOL.precision)
-        result["Length"] = "{:.{prec}f}".format(float(self._instance.length), prec=TOL.precision)
-        result["Width"] = "{:.{prec}f}".format(float(self._instance.width), prec=TOL.precision)
-        result["Depth"] = "{:.{prec}f}".format(float(self._instance.depth), prec=TOL.precision)
-        result["LeadAngleParallel"] = "yes" if self._instance.lead_angle_parallel else "no"
-        result["LeadAngle"] = "{:.{prec}f}".format(float(self._instance.lead_angle), prec=TOL.precision)
-        result["LeadInclinationParallel"] = "yes" if self._instance.lead_inclination_parallel else "no"
-        result["LeadInclination"] = "{:.{prec}f}".format(float(self._instance.lead_inclination), prec=TOL.precision)
-        result["MachiningLimits"] = {key: "yes" if value else "no" for key, value in self._instance.machining_limits.items()}
-        return result
+    @property
+    def attribute_map(self):
+        return {
+            "Orientation": "orientation",
+            "StartX": "start_x",
+            "StartY": "start_y",
+            "Angle": "angle",
+            "Inclination": "inclination",
+            "Slope": "slope",
+            "Length": "length",
+            "Width": "width",
+            "Depth": "depth",
+            "LeadAngleParallel": "lead_angle_parallel",
+            "LeadAngle": "lead_angle",
+            "LeadInclinationParallel": "lead_inclination_parallel",
+            "LeadInclination": "lead_inclination",
+            "MachiningLimits": "machining_limits",
+        }
 
 
 class LapProxy(object):
@@ -940,7 +928,7 @@ class LapProxy(object):
             The volume of the lap. Must have 6 faces.
         beam : :class:`~compas_timber.elements.Beam`
             The beam that is cut by this instance.
-        machining_limits : dict, optional
+        machining_limits : :class:`compas_timber.fabrication.MachiningLimits()` or dict
             The machining limits for the cut. Default is None.
         ref_side_index : int, optional
             The index of the reference side of the element. Default is 0.