+++ src/compas_timber/fabrication/step_joint.py
@@ -1,5 +1,4 @@
 import math
-from collections import OrderedDict
 
 from compas.geometry import Box
 from compas.geometry import Brep
@@ -12,7 +11,6 @@ from compas.geometry import angle_vectors_signed
 from compas.geometry import distance_point_point
 from compas.geometry import intersection_line_plane
 from compas.geometry import is_point_behind_plane
-from compas.tolerance import TOL
 
 from compas_timber.errors import FeatureApplicationError
 from compas_timber.utils import planar_surface_point_at
@@ -713,23 +711,16 @@ class StepJointParams(BTLxProcessingParams):
         # type: (StepJoint) -> None
         super(StepJointParams, self).__init__(instance)
 
-    def as_dict(self):
-        """Returns the parameters of the Step Joint feature as a dictionary.
-
-        Returns
-        -------
-        dict
-            The parameters of the Step Joint as a dictionary.
-        """
-        # type: () -> OrderedDict
-        result = OrderedDict()
-        result["Orientation"] = self._instance.orientation
-        result["StartX"] = "{:.{prec}f}".format(float(self._instance.start_x), prec=TOL.precision)
-        result["StrutInclination"] = "{:.{prec}f}".format(float(self._instance.strut_inclination), prec=TOL.precision)
-        result["StepDepth"] = "{:.{prec}f}".format(float(self._instance.step_depth), prec=TOL.precision)
-        result["HeelDepth"] = "{:.{prec}f}".format(float(self._instance.heel_depth), prec=TOL.precision)
-        result["StepShape"] = self._instance.step_shape
-        result["Tenon"] = "yes" if self._instance.tenon else "no"
-        result["TenonWidth"] = "{:.{prec}f}".format(float(self._instance.tenon_width), prec=TOL.precision)
-        result["TenonHeight"] = "{:.{prec}f}".format(float(self._instance.tenon_height), prec=TOL.precision)
-        return result
+    @property
+    def attribute_map(self):
+        return {
+            "Orientation": "orientation",
+            "StartX": "start_x",
+            "StrutInclination": "strut_inclination",
+            "StepDepth": "step_depth",
+            "HeelDepth": "heel_depth",
+            "StepShape": "step_shape",
+            "Tenon": "tenon",
+            "TenonWidth": "tenon_width",
+            "TenonHeight": "tenon_height",
+        }